Exercise: Create a set
By completing this exercise, you demonstrated your ability to store unique
items in a set.
You should now also be able to:

Use the union() method to combine sets into one,
Use the remove() method to remove items from a set,
Use the count property to check the number of items in a set, and
Use the insert() method to expand the set by adding a new item.

https://www.coursera.org/learn/advanced-programming-in-swift/supplement/WfTEV/exercise-create-a-set