Use map, filter and reduce to calculate values
In this exercise, you practiced using the map function of the Array class
to convert an array of objects into an array of different objects. You
also explored how to use the filter function to create an array of objects
containing only some of the values from the original array, depending on
the condition provided. Lastly, you practiced using the reduce function
call on the array of objects, to produce a single object from the
combination of all the array elements.

https://www.coursera.org/learn/advanced-programming-in-swift/supplement/7NLGN/exercise-use-map-filter-and-reduce-to-calculate-values