ForEach
In this reading, you learned that instead of using a for-in loop to
iterate over arrays or dictionaries you can use a forEach higher-order
function. You have also learned about the benefits of using forEach but
should also be aware that, unlike for-in loops, elements cannot be skipped
while using forEach.

You also explored how a for-in loop could be replaced by the forEach
function, chaining functions calls and writing code that’s more succinct.

https://www.coursera.org/learn/advanced-programming-in-swift/supplement/PuCfX/foreach