Exercise: Implement a unit test
In this exercise, you practiced how to write your own unit test. You also
created the test data, which you used to call the function that you were
testing. You then confirmed the result by using the XCTAssertEqual
function that compared the output of the function to test the expected
value.

https://www.coursera.org/learn/advanced-programming-in-swift/supplement/YTtK7/exercise-implement-a-unit-test